+++ src/main/java/frc/robot/RobotContainer.java
@@ -25,7 +25,6 @@ import frc.robot.auto.PIDAlignCommand;
 import frc.robot.constants.Constants.ControllerConstants;
 import frc.robot.constants.Constants.ScoringConstants;
 import frc.robot.constants.Constants.ShuffleboardTabNames;
-import frc.robot.constants.Constants.StatusColors;
 import frc.robot.coral.CoralSubsystem;
 import frc.robot.coral.IntakeCoralCommand;
 import frc.robot.coral.OuttakeCoralCommand;
@@ -36,6 +35,7 @@ import frc.robot.swerve.SwerveSubsystem;
 import frc.robot.swerve.SwerveTelemetry;
 import frc.robot.swerve.TunerConstants;
 import frc.robot.led.LEDSubsystem;
+import frc.robot.led.StatusColors;
 import frc.robot.utilities.CommandGenerators;
 import frc.robot.vision.VisionSubsystem;
