+++ src/main/java/frc/robot/auto/PIDAlignCommand.java
@@ -18,7 +18,7 @@ import edu.wpi.first.math.kinematics.ChassisSpeeds;
 import edu.wpi.first.math.util.Units;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.constants.Constants.AligningConstants;
-import frc.robot.constants.Constants.StatusColors;
+import frc.robot.led.StatusColors;
 import frc.robot.led.LEDSubsystem;
 import frc.robot.swerve.SwerveSubsystem;
 import frc.robot.vision.VisionSubsystem;
