+++ src/main/java/frc/robot/coral/IntakeCoralCommand.java
@@ -5,7 +5,7 @@
 package frc.robot.coral;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.constants.Constants.StatusColors;
+import frc.robot.led.StatusColors;
 import frc.robot.led.LEDSubsystem;
 
 /** A command that intakes coral and stops when it reaches the end of the intake. */
