+++ src/main/java/frc/robot/coral/OuttakeCoralCommand.java
@@ -6,7 +6,7 @@ package frc.robot.coral;
 
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.constants.Constants.StatusColors;
+import frc.robot.led.StatusColors;
 import frc.robot.led.LEDSubsystem;
 
 /** A command that ejects the coral and stops a short time after. */
