+++ src/main/java/frc/robot/elevator/ZeroElevatorCommand.java
@@ -2,7 +2,7 @@ package frc.robot.elevator;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.constants.Constants.ScoringConstants;
-import frc.robot.constants.Constants.StatusColors;
+import frc.robot.led.StatusColors;
 import frc.robot.led.LEDSubsystem;
 
 /** A command which lifts the elevator and lets it fall down slowly to reset position. */
